Refactor WifiBox component: extract WifiRow for better modularity and maintainability
